Replace Tarski option to show authors and associated functions with new is_multi_author function.
This commit deprecates the following functions:

* tarski_count_authors
* tarski_should_show_authors
* tarski_resave_show_authors

The Tarski show_authors option is now always set to true, but is never used; it should now be considered a dummy value included only for legacy reasons. The tarski_resave_show_authors is now no longer called when posts are saved or deleted.